+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [unreleased]
+
+### Changed
+- Test factories and other test files are now included in the package installation. Factories can be useful when creating project tests.
+
 ## [0.3.1] - 2016-04-13
 
 ### Added